Add level 4. Add a lot more possible scenetimeline actions. Add breakpoints to the grammar to help with debugging.

+++ core/src/com/lyeeedar/Components/StatisticsComponent.kt
@@ -130,7 +130,7 @@ class StatisticsComponent: AbstractComponent()
 		if (deathEl != null) deathEffect = AssetManager.loadParticleEffect(deathEl)
 	}
 
-	fun dealDamage(amount: Int, element: ElementType, elementalConversion: Float)
+	fun dealDamage(amount: Float, element: ElementType, elementalConversion: Float)
 	{
 		var elementalDam = (amount * elementalConversion).ciel()
 		val baseDam = amount - elementalDam
